task_backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
openai
python-dotenv
faiss-cpu
beautifulsoup4
requests